+++ src/main/resources/bootstrap/fonts/glyphicons-halflings-regular.woff2
