+++ src/main/resources/bootstrap/js/bootstrap.js
