+++ src/main/resources/bootstrap/js/npm.js
