+++ src/test/java/nl/avans/facturatie/domain/TreatmentTest.java
